+++ src/core/packetgen/utils.go
